tests/wpt/web-platform-tests/service-workers/service-worker/worker-interception.https.html

<!DOCTYPE html>
<title>Service Worker: intercepting Worker script loads</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<body>
<script>

async function setup_service_worker(t, script_url, scope) {
  const r =
      await service_worker_unregister_and_register(t, script_url, scope);
  t.add_cleanup(() => service_worker_unregister(t, scope));
  await wait_for_state(t, r.installing, 'activated');
}

promise_test(async t => {
  const worker_url = 'resources/dummy-synthesized-worker.js?dedicated';
  const service_worker = 'resources/dummy-worker-interceptor.js';
  const scope = worker_url;

  await setup_service_worker(t, service_worker, scope);
  const w = new Worker(worker_url);
  const data = await new Promise((resolve, reject) => {
    w.onmessage = e => resolve(e.data);
    w.onerror = e => reject(e.message);
  });
  assert_equals(data, 'worker loading intercepted by service worker');
}, `Verify a dedicated worker script request issued from a uncontrolled ` +
   `document is intercepted by worker's own service worker.`);

promise_test(async t => {
  const worker_url = 'resources/dummy-synthesized-worker.js?shared';
  const service_worker = 'resources/dummy-worker-interceptor.js';
  const scope = worker_url;

  await setup_service_worker(t, service_worker, scope);
  const w = new SharedWorker(worker_url);
  const data = await new Promise((resolve, reject) => {
    w.port.onmessage = e => resolve(e.data);
    w.onerror = e => reject(e.message);
  });
  assert_equals(data, 'worker loading intercepted by service worker');
}, `Verify a shared worker script request issued from a uncontrolled ` +
   `document is intercepted by worker's own service worker.`);

promise_test(async t => {
  const worker_url = 'resources/dummy-same-origin-worker.js?dedicated';
  const service_worker = 'resources/dummy-worker-interceptor.js';
  const scope = worker_url;

  await setup_service_worker(t, service_worker, scope);
  const w = new Worker(worker_url);
  const data = await new Promise((resolve, reject) => {
    w.onmessage = e => resolve(e.data);
    w.onerror = e => reject(e.message);
  });
  assert_equals(data, 'dedicated worker script loaded');
}, 'Verify a same-origin worker script served by a service worker succeeds ' +
   'in starting a dedicated worker.');

promise_test(async t => {
  const worker_url = 'resources/dummy-same-origin-worker.js?shared';
  const service_worker = 'resources/dummy-worker-interceptor.js';
  const scope = worker_url;

  await setup_service_worker(t, service_worker, scope);
  const w = new SharedWorker(worker_url);
  const data = await new Promise((resolve, reject) => {
    w.port.onmessage = e => resolve(e.data);
    w.onerror = e => reject(e.message);
  });
  assert_equals(data, 'shared worker script loaded');
}, 'Verify a same-origin worker script served by a service worker succeeds ' +
   'in starting a shared worker.');

promise_test(async t => {
  const worker_url = 'resources/dummy-cors-worker.js?dedicated';
  const service_worker = 'resources/dummy-worker-interceptor.js';
  const scope = worker_url;

  await setup_service_worker(t, service_worker, scope);
  const w = new Worker(worker_url);
  const watcher = new EventWatcher(t, w, ['message', 'error']);
  await watcher.wait_for('error');
}, 'Verify a cors worker script served by a service worker fails dedicated ' +
   'worker start.');

promise_test(async t => {
  const worker_url = 'resources/dummy-cors-worker.js?shared';
  const service_worker = 'resources/dummy-worker-interceptor.js';
  const scope = worker_url;

  await setup_service_worker(t, service_worker, scope);
  const w = new SharedWorker(worker_url);
  const watcher = new EventWatcher(t, w, ['message', 'error']);
  await watcher.wait_for('error');
}, 'Verify a cors worker script served by a service worker fails shared ' +
   'worker start.');

promise_test(async t => {
  const worker_url = 'resources/dummy-no-cors-worker.js?dedicated';
  const service_worker = 'resources/dummy-worker-interceptor.js';
  const scope = worker_url;

  await setup_service_worker(t, service_worker, scope);
  const w = new Worker(worker_url);
  const watcher = new EventWatcher(t, w, ['message', 'error']);
  await watcher.wait_for('error');
}, 'Verify a no-cors cross-origin worker script served by a service worker ' +
   'fails dedicated worker start.');

promise_test(async t => {
  const worker_url = 'resources/dummy-no-cors-worker.js?shared';
  const service_worker = 'resources/dummy-worker-interceptor.js';
  const scope = worker_url;

  await setup_service_worker(t, service_worker, scope);
  const w = new SharedWorker(worker_url);
  const watcher = new EventWatcher(t, w, ['message', 'error']);
  await watcher.wait_for('error');
}, 'Verify a no-cors cross-origin worker script served by a service worker ' +
   'fails shared worker start.');

promise_test(async t => {
  const subdoc_url = 'resources/worker-interception-iframe.https.html';
  const service_worker = 'resources/worker-load-interceptor.js';
  const scope = 'resources/load_worker.js?dedicated';

  await setup_service_worker(t, service_worker, scope);
  const frame = await with_iframe(subdoc_url);
  t.add_cleanup(() => frame.remove());
  const data = await new Promise((resolve, reject) => {
    const channel = new MessageChannel();
    channel.port1.onmessage = e => resolve(e.data);
    frame.contentWindow.postMessage(
        { script: 'load_worker.js?dedicated', type: 'worker' },
        '*', [channel.port2]);
  });
  assert_equals(data.results, 'finish');
}, 'Verify subresource requests on a dedicated worker controlled by a ' +
   'service worker.');

promise_test(async t => {
  const subdoc_url = 'resources/worker-interception-iframe.https.html';
  const service_worker = 'resources/worker-load-interceptor.js';
  const scope = 'resources/load_worker.js?shared';

  await setup_service_worker(t, service_worker, scope);
  const frame = await with_iframe(subdoc_url);
  t.add_cleanup(() => frame.remove());
  const data = await new Promise((resolve, reject) => {
    const channel = new MessageChannel();
    channel.port1.onmessage = e => resolve(e.data);
    frame.contentWindow.postMessage(
        { script: 'load_worker.js?shared', type: 'sharedworker' },
        '*', [channel.port2]);
  });
  assert_equals(data.results, 'finish');
}, 'Verify subresource requests on a shared worker controlled by a service ' +
   'worker.');

promise_test(async t => {
  const subdoc_url = 'resources/worker-interception-iframe.https.html';
  const service_worker = 'resources/worker-load-interceptor.js';
  const scope = 'resources/nested_load_worker.js?dedicated';

  await setup_service_worker(t, service_worker, scope);
  const frame = await with_iframe(subdoc_url);
  t.add_cleanup(() => frame.remove());
  const data = await new Promise((resolve, reject) => {
    const channel = new MessageChannel();
    channel.port1.onmessage = e => resolve(e.data);
    frame.contentWindow.postMessage(
        { script: 'nested_load_worker.js?dedicated', type: 'worker' },
        '*', [channel.port2]);
  });
  assert_equals(data.results, 'finish');
}, 'Verify subresource requests on a nested dedicated worker created from a ' +
   'dedicated worker and controlled by a service worker.');

promise_test(async t => {
  const subdoc_url = 'resources/worker-interception-iframe.https.html';
  const service_worker = 'resources/worker-load-interceptor.js';
  const scope = 'resources/nested_load_worker.js?shared';

  await setup_service_worker(t, service_worker, scope);
  const frame = await with_iframe(subdoc_url);
  t.add_cleanup(() => frame.remove());
  const data = await new Promise((resolve, reject) => {
    const channel = new MessageChannel();
    channel.port1.onmessage = e => resolve(e.data);
    frame.contentWindow.postMessage(
        { script: 'nested_load_worker.js?shared', type: 'sharedworker' },
        '*', [channel.port2]);
  });
  assert_equals(data.results, 'finish');
}, 'Verify subresource requests on a nested dedicated worker created from a ' +
   'shared worker and controlled by a service worker.');

</script>
</body>
